Rename src/auth -> src/handshake
